Maven tests successful, completed BasicStockService factory implementation.
