+++ Week3/src/main/java/edu/bradyreed/advjava/Week3/IStockService.java
@@ -1,0 +1,9 @@
+package edu.bradyreed.advjava.Week3;
+
+// Implemented by BasicStockService.java
+
+public interface IStockService {
+	
+	//Returns the value for a specified stock symbol
+	StockQuote getQuote(String symbol);
+}
