--- Week3/src/main/java/edu/bradyreed/advjava/Week3/factories/StockServiceFactory.java
@@ -1,5 +1,0 @@
-package edu.bradyreed.advjava.Week3.factories;
-
-public class StockServiceFactory {
-
-}
