--- Week3/src/main/java/edu/bradyreed/advjava/Week3/interfaces/IStockService.java
@@ -1,8 +1,0 @@
-package edu.bradyreed.advjava.Week3.interfaces;
-
-import edu.bradyreed.advjava.Week3.classes.StockQuote;
-
-public interface IStockService {
-	
-	StockQuote getQuote(String symbol);
-}
