--- Week3/src/test/java/edu/bradyreed/advjava/Week3/AppTest.java
@@ -1,20 +1,0 @@
-package edu.bradyreed.advjava.Week3;
-
-import static org.junit.Assert.assertTrue;
-
-import org.junit.Test;
-
-/**
- * Unit test for simple App.
- */
-public class AppTest 
-{
-    /**
-     * Rigorous Test :-)
-     */
-    @Test
-    public void shouldAnswerWithTrue()
-    {
-        assertTrue( true );
-    }
-}
